Created new folder for Trip form to handle both new and update logic, and started building update function
