+++ src/components/NewTripForm/NewTripForm.jsx
@@ -1,68 +1,77 @@
-import { useState } from "react";
-import { useNavigate } from "react-router";
-import * as tripService from "../../services/tripService";
+//Created TripForm component to handle both creating and updating trips in the same component.
 
-const NewTripForm = () => {
-  const navigate = useNavigate();
-  const [message, setMessage] = useState("");
-  const [formData, setFormData] = useState({
-    title: "",
-    description: "",
-    // destination: "", 
-  });
 
-  const handleChange = (event) => {
-    setMessage("");
-    setFormData({ ...formData, [event.target.name]: event.target.value });
-  };
+// import { useState } from "react";
+// import { useNavigate } from "react-router";
+// import * as tripService from "../../services/tripService";
 
-  const handleSubmit = async (event) => {
-    event.preventDefault();
-    try {
-      await tripService.create(formData);
-      setMessage("Trip created successfully!");
-      navigate("/trips");
-    } catch (error) {
-      console.log(error);
-      setMessage("Error creating trip. Please try again.");
-    }
-  };
+// const NewTripForm = () => {
+//   const navigate = useNavigate();
+//   const [message, setMessage] = useState("");
+//   const [formData, setFormData] = useState({
+//     title: "",
+//     description: "",
+//     // destination: "", 
+//   });
 
-  return (
-    <main>
-      <h1>Create a new Trip</h1>
+//   const handleChange = (event) => {
+//     setMessage("");
+//     setFormData({ ...formData, [event.target.name]: event.target.value });
+//   };
 
-      <form onSubmit={handleSubmit}>
-        <label htmlFor="title">Title:</label>
-        <input
-          type="text"
-          name="title"
-          id="title"
-          value={formData.title}
-          onChange={handleChange}
-        />
+//   const handleSubmit = async (event) => {
+//     event.preventDefault();
+//     try {
+//       await tripService.create(formData);
+//       setMessage("Trip created successfully!");
+//       navigate("/trips");
+//     } catch (error) {
+//       console.log(error);
+//       setMessage("Error creating trip. Please try again.");
+//     }
+//   };
 
-        <label htmlFor="description">Description:</label>
-        <textarea
-          name="description"
-          id="description"
-          value={formData.description}
-          onChange={handleChange}
-        ></textarea>
+//   const handleUpdateTrip = async (tripId, newTripFormData) => {
+//     const updatedTrip = await tripService.update(tripId, newTripFormData);
+//     setTrip(tripService.map((trip) => (trip._id === tripId ? updatedTrip : trip)));
+//     navigate(`/trips/${tripId}`);
+//   };
 
-        {/* <label htmlFor="destination">Destination:</label>
-        <input
-          type="text"
-          name="destination"
-          id="destination"
-          value={formData.destination}
-          onChange={handleChange}
-        /> */}
+//   return (
+//     <main>
+//       <h1>Create a new Trip</h1>
 
-        <button type="submit">Create Trip</button>
-      </form>
-    </main>
-  );
-};
+//       <form onSubmit={handleSubmit}>
+//         <label htmlFor="title">Title:</label>
+//         <input
+//           type="text"
+//           name="title"
+//           id="title"
+//           value={formData.title}
+//           onChange={handleChange}
+//         />
 
-export default NewTripForm;
+//         <label htmlFor="description">Description:</label>
+//         <textarea
+//           name="description"
+//           id="description"
+//           value={formData.description}
+//           onChange={handleChange}
+//         ></textarea>
+
+//         {/* <label htmlFor="destination">Destination:</label>
+//         <input
+//           type="text"
+//           name="destination"
+//           id="destination"
+//           value={formData.destination}
+//           onChange={handleChange}
+//         /> */}
+
+//         <button type="submit">Create Trip</button>
+//       </form>
+//     </main>
+//   );
+// };
+
+// export default NewTripForm;
